test_forwardmsg_hashing, test_forwardmsg_cacheable_flag, test_duplicate_forwardmsg_caching
